refactor(runtime): move VRAMBudget into vram_budget.h, drop executor.h include
