+++ src/exec/executor.h
@@ -21,6 +21,7 @@
 #include "exec/quant_scratch.h"
 #include "exec/quant_pipeline.h"
 #include "runtime/storage_planner.h"
+#include "runtime/vram_budget.h"
 #include "runtime/config.h"
 #include <cuda_runtime.h>
 #include <cuda_fp16.h>
@@ -33,19 +34,6 @@ namespace imp {
 
 class LoraAdapter;
 struct LoraWeights;
-
-// VRAM budget for weight cache allocation (computed by Engine::plan_vram_budget).
-// Replaces ad-hoc "remaining_budget" with per-phase caps computed upfront.
-struct VRAMBudget {
-    enum Strategy { FP8_PREFILL_NVFP4_DECODE, NVFP4_DECODE_ONLY, FP16_ONLY };
-    Strategy strategy = FP16_ONLY;
-    size_t kv_cache_bytes = 0;
-    size_t fp8_cache_bytes = 0;  // 0 for sub-8-bit models
-    size_t nvfp4_cache_bytes = 0;
-    size_t reserve_bytes = 1024ULL * 1024 * 1024;  // 1 GiB safety
-    int kv_max_blocks = 0;
-    bool nvfp4_second_pass = false;  // true → re-run NVFP4 after FP16-Free
-};
 
 // Nvfp4DecodeContext moved to exec/quant_pipeline.h (build-only; consumed by
 // the QuantPipeline phase-3 helpers).
